Add basic recipe for OpenFOAM-8 (just MC on Eiger) (#60)
adds a recipe for creating an image containing OpenFOAM Foundation v8 on Eiger, that:
- builds `openfoam-org@8`
- provides module `openfoam/8`
- provides two views, `openfoam` for the average user that just need the package, and `develop` for users that needs dependencies to build `openfoam-org` from scratch on their own

note: `openfoam-org@8` provided by the spack package is not the very last tagged version on the GitHub repo. I already played with it and I might have a solution to adapt this recipe to actually build a custom tag. But I will take care of it in a separate PR.