fix cp2k/2024.1 COSMA support on gh200  (#107)
A patch was added in #95 for `COSTA` and `COSMA`, but it was only applied to `COSMA`. This prevented to build with `-DCP2K_USE_COSMA=ON` from an view. This PR actually applies the provided patch.

This issue was likely not spotted because of a typo in the CMake command I used, fixed in #106.

The patches have been contributed upstream:
* https://github.com/eth-cscs/COSMA/pull/144
* https://github.com/eth-cscs/COSTA/pull/22