The package is a cmake package
deepmd is a python package with a c++ library that is need for other
external packages such as lammps. Unfortunately `pip install` screws this up
and we have to manually install the library to get thing work as expected.

So the package is now a cmake package and treat the python part as an extension.