Don't exclude gcc module in `prgenv-gnu` (#247)
When going from prgenv-gnu/24.11 to 25.6, I blindly updated the
`exclude` field for modules to exclude the newer system compiler. After
the bootstrapping was removed, this excludes the actual compiler we want
to expose as a module. I think we can get away without excluding any
specific modules. Only the correct `gcc` is considered at least in the
case of prgenv-gnu/25.6 even when the exclude field is empty.