Add linalg/linalg-complex 25.10 (#255)
Updated all four linalg uenvs to an environment similar to the latest
prgenv-gnu, i.e. using CUDA 12.9, GCC 14, cray-mpich 8.1.32 etc. I had
to pin some packages, but added comments with links to issues for them.
This is using a recent commit from spack-packages (instead of 2025.07)
to get a version of the superlu-dist package which applies a required
patch.