Add linalg recipe for beverin (#262)
Disables kokkos and related packages (`+kokkos` variants and trilinos)
because kokkos only wants one architecture at a time, and I currently
don't have a need or motivation to set that up e.g. in two separate
uenvs. If there are requests I can try to update that in a future
iteration. This is mostly useful for dlaf benchmarks on beverin and the
other packages come as a bonus.