Add rccl and rccl-tests to rocm uenv (#271)
Requires https://github.com/eth-cscs/stackinator/pull/268 and
https://github.com/eth-cscs/stackinator/pull/269.

With the `full` concretizer strategy, unfortunately some packages are by
default concretized with `llvm-amdgpu` as compiler. I've forced them to
use `gcc` as the compiler.

`rccl-tests` and `aws-ofi-rccl` are unversioned, so I'm pinning them to
recent commits just for reproduciblity. I need to try
https://github.com/HewlettPackard/open-ofi-xccl eventually in place of
`aws-ofi-rccl` as `open-ofi-xccl` should be the preferred long term
solution as far as I understand, but for now `aws-ofi-rccl` seems to
work and gives decent performance.